pyproject.toml
[tool.poetry]
name = "soulaibot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
openai = "^0.27.0"
pydantic = "^1.10.5"
gunicorn = "^20.1.0"
python-telegram-bot = "^20.1"
tiktoken = "^0.3.1"
backoff = "^2.2.1"
google-cloud-datastore = "^2.15.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
